Revision: pac--0.1--patch-5 Archive: stig@brautaset.org--private Creator: Stig Brautaset <stig@brautaset.org> Date: Wed Jun 12 11:47:32 BST 2002 Standard-date: 2002-06-12 10:47:32 GMT Summary: warp tunnel working! Keywords: New-files: {arch}/pac/pac/pac--0.1/stig@brautaset.org--private/patch-log/patch-5   src/.arch-ids/default.h.id .arch-ids/NEWS.id src/default.h   NEWS Modified-files: Makefile.am TODO configure.in src/creature_common.c   src/creature_common.h src/creature_ghosts.c   src/creature_pac.c src/gui.c src/gui.h src/gui_unix_ncurses.c   src/gui_unix_ncurses.h src/layout.dat src/pac.c src/pac.h   src/score.c New-patches: stig@brautaset.org--private/pac--0.1--patch-5
I thought it was a major problem to get the tunnel working, but it seems
is was not that hard. Using addch() instead of printw() to draw the
creatures and introducing a bit of trickery with the modulus operator
and voilá :)

A side effect is that I have to use more basic ascii graphics to draw
the creatures, because the ncurses addch() does not play well with
characters > 127 whereas printw() would.