Revision: pac--0.2--patch-1 Archive: stig@brautaset.org--private Creator: Stig Brautaset <stig@brautaset.org> Date: Sat Jun 22 12:55:32 BST 2002 Standard-date: 2002-06-22 11:55:32 GMT Summary: cleaned up configure.in/bootstrap + myrand() Keywords: New-files: {arch}/pac/pac/pac--0.2/stig@brautaset.org--private/patch-log/patch-1 Removed-files: src/.arch-ids/default.h.id src/default.h Modified-files: bootstrap configure.in src/Makefile.am src/creature_ghosts.c   src/creature_ghosts.h src/gui_unix_ncurses.c   src/gui_unix_ncurses.h src/pac.c src/pac.h New-patches: stig@brautaset.org--private/pac--0.2--patch-1
Cleaned up configure.in/bootstrap and Makefile.am a bit. It now work the
way it was intended, and I understand autoconf/automake better.

Made myrand() into a macro instead of the function it was. We lose the
overhead of a function call, but I'm not sure we gain much.

Starting to add inclusion guards for some constructs.